+++ sympy/geometry/line.py
@@ -36,6 +36,7 @@ from sympy.logic.boolalg import And
 from sympy.matrices import Matrix
 from sympy.sets.sets import Intersection
 from sympy.simplify.simplify import simplify
+from sympy.solvers.solvers import solve
 from sympy.solvers.solveset import linear_coeffs
 from sympy.utilities.exceptions import SymPyDeprecationWarning
 from sympy.utilities.misc import Undecidable, filldedent
@@ -2556,7 +2557,6 @@ class Line3D(LinearEntity3D, Line):
                 feature="equation() no longer needs 'k'",
                 issue=13742,
                 deprecated_since_version="1.2").warn()
-        from sympy.solvers.solvers import solve
         x, y, z, k = [_symbol(i, real=True) for i in (x, y, z, 'k')]
         p1, p2 = self.points
         d1, d2, d3 = p1.direction_ratio(p2)
