+++ sympy/integrals/meijerint.py
@@ -64,9 +64,6 @@ from sympy.functions.special.singularity_functions import SingularityFunction
 from .integrals import Integral
 from sympy.logic.boolalg import And, Or, BooleanAtom, Not, BooleanFunction
 from sympy.polys import cancel, factor
-from sympy.simplify.fu import sincos_to_sum
-from sympy.simplify import (collect, gammasimp, hyperexpand, powdenest,
-                            powsimp, simplify)
 from sympy.utilities.iterables import multiset_partitions
 from sympy.utilities.misc import debug as _debug
 
@@ -337,6 +334,7 @@ def _get_coeff_exp(expr, x):
     >>> _get_coeff_exp(x**3, x)
     (1, 3)
     """
+    from sympy.simplify import powsimp
     (c, m) = expand_power_base(powsimp(expr)).as_coeff_mul(x)
     if not m:
         return c, S.Zero
@@ -900,6 +898,7 @@ def _int0oo_1(g, x):
     >>> _int0oo_1(meijerg([a], [b], [c], [d], x*y), x)
     gamma(-a)*gamma(c + 1)/(y*gamma(-d)*gamma(b + 1))
     """
+    from sympy.simplify import gammasimp
     # See [L, section 5.6.1]. Note that s=1.
     eta, _ = _get_coeff_exp(g.argument, x)
     res = 1/eta
@@ -984,6 +983,7 @@ def _rewrite_saxena(fac, po, g1, g2, x, full_pb=False):
     g1 = meijerg(tr(g1.an), tr(g1.aother), tr(g1.bm), tr(g1.bother), a1*x)
     g2 = meijerg(g2.an, g2.aother, g2.bm, g2.bother, a2*x)
 
+    from sympy.simplify import powdenest
     return powdenest(fac, polar=True), g1, g2
 
 
@@ -1328,6 +1328,7 @@ def _rewrite_inversion(fac, po, g, x):
 
     def tr(l):
         return [t + s/b for t in l]
+    from sympy.simplify import powdenest
     return (powdenest(fac/a**(s/b), polar=True),
             meijerg(tr(g.an), tr(g.aother), tr(g.bm), tr(g.bother), g.argument))
 
@@ -1558,6 +1559,7 @@ def _rewrite_single(f, x, recursive=True):
             return inverse_mellin_transform(F, s, x, strip,
                                             as_meijerg=True, needeval=True)
         except MellinTransformStripError:
+            from sympy.simplify import simplify
             return inverse_mellin_transform(
                 simplify(cancel(expand(F))), s, x, strip,
                 as_meijerg=True, needeval=True)
@@ -1568,6 +1570,7 @@ def _rewrite_single(f, x, recursive=True):
     def my_integrator(f, x):
         r = _meijerint_definite_4(f, x, only_double=True)
         if r is not None:
+            from sympy.simplify import hyperexpand
             res, cond = r
             res = _my_unpolarify(hyperexpand(res, rewrite='nonrepsmall'))
             return Piecewise((res, cond),
@@ -1684,6 +1687,7 @@ def meijerint_indefinite(f, x):
             _rewrite_hyperbolics_as_exp(f), x)
         if rv:
             if not isinstance(rv, list):
+                from sympy.simplify.radsimp import collect
                 return collect(factor_terms(rv), rv.atoms(exp))
             results.extend(rv)
     if results:
@@ -1693,6 +1697,7 @@ def meijerint_indefinite(f, x):
 def _meijerint_indefinite_1(f, x):
     """ Helper that does not attempt any substitution. """
     _debug('Trying to compute the indefinite integral of', f, 'wrt', x)
+    from sympy.simplify import hyperexpand, powdenest
 
     gs = _rewrite1(f, x)
     if gs is None:
@@ -1910,6 +1915,7 @@ def meijerint_definite(f, x, a, b):
             _rewrite_hyperbolics_as_exp(f_), x_, a_, b_)
         if rv:
             if not isinstance(rv, list):
+                from sympy.simplify.radsimp import collect
                 rv = (collect(factor_terms(rv[0]), rv[0].atoms(exp)),) + rv[1:]
                 return rv
             results.extend(rv)
@@ -1940,6 +1946,7 @@ def _guess_expansion(f, x):
             saw.add(expanded)
 
     if orig.has(cos, sin):
+        from sympy.simplify.fu import sincos_to_sum
         reduced = sincos_to_sum(orig)
         if reduced not in saw:
             res += [(reduced, 'trig power reduction')]
@@ -2020,6 +2027,7 @@ def _meijerint_definite_4(f, x, only_double=False):
     The parameter ``only_double`` is used internally in the recursive algorithm
     to disable trying to rewrite f as a single G-function.
     """
+    from sympy.simplify import hyperexpand
     # This function does (2) and (3)
     _debug('Integrating', f)
     # Try single G function.
@@ -2178,6 +2186,7 @@ def meijerint_inversion(f, x, t):
             _debug('But cond is always False.')
         else:
             _debug('Result before branch substitution:', res)
+            from sympy.simplify import hyperexpand
             res = _my_unpolarify(hyperexpand(res))
             if not res.has(Heaviside):
                 res *= Heaviside(t)
