+++ sympy/polys/polyroots.py
@@ -25,7 +25,6 @@ from sympy.polys.polyquinticconst import PolyQuintic
 from sympy.polys.polytools import Poly, cancel, factor, gcd_list, discriminant
 from sympy.polys.rationaltools import together
 from sympy.polys.specialpolys import cyclotomic_poly
-from sympy.simplify.simplify import simplify, powsimp
 from sympy.utilities import public
 
 
@@ -38,6 +37,7 @@ def roots_linear(f):
         if dom.is_Composite:
             r = factor(r)
         else:
+            from sympy.simplify.simplify import simplify
             r = simplify(r)
 
     return [r]
@@ -75,6 +75,7 @@ def roots_quadratic(f):
         if dom.is_Composite:
             return factor(expr)
         else:
+            from sympy.simplify.simplify import simplify
             return simplify(expr)
 
     if c is S.Zero:
@@ -653,6 +654,7 @@ def roots_quintic(f):
 
 
 def _quintic_simplify(expr):
+    from sympy.simplify.simplify import powsimp
     expr = powsimp(expr)
     expr = cancel(expr)
     return together(expr)
