+++ sympy/series/limits.py
@@ -8,9 +8,6 @@ from sympy.functions.elementary.exponential import (exp, log)
 from sympy.functions.special.gamma_functions import gamma
 from sympy.polys import PolynomialError, factor
 from sympy.series.order import Order
-from sympy.simplify.powsimp import powsimp
-from sympy.simplify.ratsimp import ratsimp
-from sympy.simplify.simplify import nsimplify, together
 from .gruntz import gruntz
 
 def limit(e, z, z0, dir="+"):
@@ -82,6 +79,7 @@ def heuristics(e, z, z0, dir):
             return
     elif e.is_Mul or e.is_Add or e.is_Pow or e.is_Function:
         r = []
+        from sympy.simplify.simplify import together
         for a in e.args:
             l = limit(a, z, z0, dir)
             if l.has(S.Infinity) and l.is_finite is None:
@@ -119,6 +117,7 @@ def heuristics(e, z, z0, dir):
 
             if rv is S.NaN:
                 try:
+                    from sympy.simplify.ratsimp import ratsimp
                     rat_e = ratsimp(e)
                 except PolynomialError:
                     return
@@ -267,6 +266,7 @@ class Limit(Expr):
             # prevent rounding errors which can lead to unexpected execution
             # of conditional blocks that work on comparisons
             # Also see comments in https://github.com/sympy/sympy/issues/19453
+            from sympy.simplify.simplify import nsimplify
             e = nsimplify(e)
         e = set_signs(e)
 
@@ -306,6 +306,7 @@ class Limit(Expr):
             coeff, ex = newe.leadterm(z, cdir=cdir)
         except (ValueError, NotImplementedError, PoleError):
             # The NotImplementedError catching is for custom functions
+            from sympy.simplify.powsimp import powsimp
             e = powsimp(e)
             if e.is_Pow:
                 r = self.pow_heuristics(e)
