+++ sympy/series/sequences.py
@@ -3,7 +3,7 @@ from sympy.core.cache import cacheit
 from sympy.core.containers import Tuple
 from sympy.core.decorators import call_highest_priority
 from sympy.core.parameters import global_parameters
-from sympy.core.function import AppliedUndef
+from sympy.core.function import AppliedUndef, expand
 from sympy.core.mul import Mul
 from sympy.core.numbers import Integer
 from sympy.core.relational import Eq
@@ -11,12 +11,11 @@ from sympy.core.singleton import S, Singleton
 from sympy.core.sorting import ordered
 from sympy.core.symbol import Dummy, Symbol, Wild
 from sympy.core.sympify import sympify
+from sympy.matrices import Matrix
 from sympy.polys import lcm, factor
 from sympy.sets.sets import Interval, Intersection
-from sympy.simplify import simplify
 from sympy.tensor.indexed import Idx
 from sympy.utilities.iterables import flatten, is_sequence, iterable
-from sympy.core.function import expand
 
 
 ###############################################################################
@@ -336,7 +335,7 @@ class SeqBase(Basic):
         >>> sequence(lucas(n)).find_linear_recurrence(15,gfvar=x)
         ([1, 1], (x - 2)/(x**2 + x - 1))
         """
-        from sympy.matrices import Matrix
+        from sympy.simplify import simplify
         x = [simplify(expand(t)) for t in self[:n]]
         lx = len(x)
         if d is None:
