+++ sympy/simplify/simplify.py
@@ -1,5 +1,7 @@
 from collections import defaultdict
 
+from sympy.concrete.products import Product
+from sympy.concrete.summations import Sum
 from sympy.core import (Basic, S, Add, Mul, Pow, Symbol, sympify,
                         expand_func, Function, Dummy, Expr, factor_terms,
                         expand_power_exp, Eq)
@@ -19,12 +21,17 @@ from sympy.functions.elementary.complexes import unpolarify, Abs, sign
 from sympy.functions.elementary.exponential import ExpBase
 from sympy.functions.elementary.hyperbolic import HyperbolicFunction
 from sympy.functions.elementary.integers import ceiling
-from sympy.functions.elementary.piecewise import Piecewise, piecewise_fold
+from sympy.functions.elementary.piecewise import (Piecewise, piecewise_fold,
+                                                  piecewise_simplify)
 from sympy.functions.elementary.trigonometric import TrigonometricFunction
 from sympy.functions.special.bessel import (BesselBase, besselj, besseli,
                                             besselk, bessely, jn)
 from sympy.functions.special.tensor_functions import KroneckerDelta
+from sympy.integrals.integrals import Integral
+from sympy.matrices.expressions import (MatrixExpr, MatAdd, MatMul,
+                                            MatPow, MatrixSymbol)
 from sympy.polys import together, cancel, factor
+from sympy.polys.numberfields.minpoly import _is_sum_surds, _minimal_polynomial_sq
 from sympy.simplify.combsimp import combsimp
 from sympy.simplify.cse_opts import sub_pre, sub_post
 from sympy.simplify.hyperexpand import hyperexpand
@@ -207,14 +214,6 @@ def _separatevars_dict(expr, symbols):
         ret[k] = Mul(*v)
 
     return ret
-
-
-def _is_sum_surds(p):
-    args = p.args if p.is_Add else [p]
-    for y in args:
-        if not ((y**2).is_Rational and y.is_extended_real):
-            return False
-    return True
 
 
 def posify(eq):
@@ -675,7 +674,6 @@ def simplify(expr, ratio=1.7, measure=count_ops, rational=False, inverse=False, 
                 expr = kroneckersimp(expr)
             # Still a Piecewise?
             if expr.has(Piecewise):
-                from sympy.functions.elementary.piecewise import piecewise_simplify
                 # Do not apply doit on the segments as it has already
                 # been done above, but simplify
                 expr = piecewise_simplify(expr, deep=True, doit=False)
@@ -707,10 +705,6 @@ def simplify(expr, ratio=1.7, measure=count_ops, rational=False, inverse=False, 
         # expression with gamma functions or non-integer arguments is
         # automatically passed to gammasimp
         expr = combsimp(expr)
-
-    from sympy.concrete.products import Product
-    from sympy.concrete.summations import Sum
-    from sympy.integrals.integrals import Integral
 
     if expr.has(Sum):
         expr = sum_simplify(expr, **kwargs)
@@ -768,8 +762,6 @@ def simplify(expr, ratio=1.7, measure=count_ops, rational=False, inverse=False, 
 
 def sum_simplify(s, **kwargs):
     """Main function for Sum simplification"""
-    from sympy.concrete.summations import Sum
-
     if not isinstance(s, Add):
         s = s.xreplace({a: sum_simplify(a, **kwargs)
             for a in s.atoms(Add) if a.has(Sum)})
@@ -801,8 +793,6 @@ def sum_combine(s_t):
        Attempts to simplify a list of sums, by combining limits / sum function's
        returns the simplified sum
     """
-    from sympy.concrete.summations import Sum
-
     used = [False] * len(s_t)
 
     for method in range(2):
@@ -843,7 +833,6 @@ def factor_sum(self, limits=None, radical=False, clear=False, fraction=False, si
     y*Sum(x, (x, 1, 3))
     """
     # XXX deprecate in favor of direct call to factor_terms
-    from sympy.concrete.summations import Sum
     kwargs = dict(radical=radical, clear=clear,
         fraction=fraction, sign=sign)
     expr = Sum(self, *limits) if limits else self
@@ -852,8 +841,6 @@ def factor_sum(self, limits=None, radical=False, clear=False, fraction=False, si
 
 def sum_add(self, other, method=0):
     """Helper function for Sum simplification"""
-    from sympy.concrete.summations import Sum
-
     #we know this is something in terms of a constant * a sum
     #so we temporarily put the constants inside for simplification
     #then simplify the result
@@ -898,8 +885,6 @@ def sum_add(self, other, method=0):
 
 def product_simplify(s):
     """Main function for Product simplification"""
-    from sympy.concrete.products import Product
-
     terms = Mul.make_args(s)
     p_t = [] # Product Terms
     o_t = [] # Other Terms
@@ -932,8 +917,6 @@ def product_simplify(s):
 
 def product_mul(self, other, method=0):
     """Helper function for Product simplification"""
-    from sympy.concrete.products import Product
-
     if type(self) is type(other):
         if method == 0:
             if self.limits == other.limits:
@@ -962,7 +945,6 @@ def _nthroot_solve(p, n, prec):
      helper function for ``nthroot``
      It denests ``p**Rational(1, n)`` using its minimal polynomial
     """
-    from sympy.polys.numberfields.minpoly import _minimal_polynomial_sq
     from sympy.solvers import solve
     while n % 2 == 0:
         p = sqrtdenest(sqrt(p))
@@ -1681,9 +1663,6 @@ def nc_simplify(expr, deep=True):
     (a*b*a*b)**2*(a*c)**2
 
     '''
-    from sympy.matrices.expressions import (MatrixExpr, MatAdd, MatMul,
-                                                MatPow, MatrixSymbol)
-
     if isinstance(expr, MatrixExpr):
         expr = expr.doit(inv_expand=False)
         _Add, _Mul, _Pow, _Symbol = MatAdd, MatMul, MatPow, MatrixSymbol
