+++ sympy/tensor/array/dense_ndim_array.py
@@ -5,7 +5,6 @@ from sympy.core.basic import Basic
 from sympy.core.containers import Tuple
 from sympy.core.singleton import S
 from sympy.core.sympify import _sympify
-from sympy.simplify.simplify import simplify
 from sympy.tensor.array.mutable_ndim_array import MutableNDimArray
 from sympy.tensor.array.ndim_array import NDimArray, ImmutableNDimArray, ArrayKind
 from sympy.utilities.iterables import flatten
@@ -158,6 +157,7 @@ class ImmutableDenseNDimArray(DenseNDimArray, ImmutableNDimArray): # type: ignor
         return MutableDenseNDimArray(self)
 
     def _eval_simplify(self, **kwargs):
+        from sympy.simplify.simplify import simplify
         return self.applyfunc(simplify)
 
 class MutableDenseNDimArray(DenseNDimArray, MutableNDimArray):
